Tony 2 (#16)
* Create Sollution

* Update Create.cshtml

* Create CRUD

* Bootstrap

* Pagination

* Page validation

* Session validation (host and admin), fix navbar index

* Squashed commit of the following:

commit 44e4cec269a5da1835b841ce62d8c508741becc3
Author: tonynguyen2512 <nguyenhathien2512@gmail.com>
Date:   Fri Mar 11 11:52:56 2022 +0700

    Session validation (host and admin), fix navbar index

commit 1962bc90c821dd6c057576c356a8b158041f50ca
Author: tonynguyen2512 <nguyenhathien2512@gmail.com>
Date:   Fri Mar 11 09:22:53 2022 +0700

    Page validation

commit 9ce6453223ce8dca48a0d2b07f5d9f86bf64612b
Merge: ad5a1b2 a73f49e
Author: tonynguyen2512 <nguyenhathien2512@gmail.com>
Date:   Thu Mar 10 22:39:18 2022 +0700

    Merge branch 'main' into Tony-2

commit ad5a1b252d85964c5aa6f54851c012db3f4326c5
Author: tonynguyen2512 <nguyenhathien2512@gmail.com>
Date:   Thu Mar 10 17:36:22 2022 +0700

    Pagination

commit 76458f6e4d6d576313f84965568fa0fbd61bae8b
Author: tonynguyen2512 <nguyenhathien2512@gmail.com>
Date:   Wed Mar 9 23:05:57 2022 +0700

    Bootstrap

commit acd39e797fc98361c6e8592dbae077c898d69306
Merge: f0b92d1 fbcb06b
Author: tonynguyen2512 <nguyenhathien2512@gmail.com>
Date:   Wed Mar 9 19:45:24 2022 +0700

    Merge branch 'main' into Tony-2

commit f0b92d11dcc631ea62eef989e4c493bcbbad0f5e
Author: tonynguyen2512 <nguyenhathien2512@gmail.com>
Date:   Wed Mar 9 16:29:51 2022 +0700

    Create CRUD

commit d64e89bdc207067bfd876e1abd9860cabf092879
Author: tonynguyen2512 <nguyenhathien2512@gmail.com>
Date:   Wed Mar 9 15:14:17 2022 +0700

    Update Create.cshtml

* Update UI

* Update UI guest

* Update UI ver2

* Update Event-Management-System-Project-Document-PRN221.docx

* Update Index_Guest.cshtml

* Update Guest UI

* Delete Event-Management-System-Project-Document-PRN221.docx

* Create Event-Management-System-Project-Document-PRN221.docx

* add pagination and search for history, seach for product guest

* Update Pagination

* Update Validation

+++ EMS/EMS/Pages/Event/Delete.cshtml.cs
@@ -62,6 +62,7 @@ namespace EMS.Pages.Event
 
             if (Event != null)
             {
+                if (Event.IsPublic == true) return Page();
                 _context.Events.Remove(Event);
                 await _context.SaveChangesAsync();
             }
